Create a wrapper struct for `fastcgi::Request`
Trying out the Conduit API, as it seems like nice interface that I can
plug the FastCGI server into.

To do this, I need a server type, and some way to convert a
`fastcgi::Request` into a `conduit::Request`. Doing this with a new
local type that I can use to implement `conduit::RequestExt`.