server.rs: Fix HTTP status code header
I was getting persistent 500 errors on Apache with mod_fastcgi:

    FastCGI: comm with server "/Applications/MAMP/htdocs/test.fcgi"
    aborted: error parsing headers: malformed header 'HTTP/1.1 200 OK'

Forgot that you need to use a "Status" header for FastCGI to set the
HTTP status code, like:

    Status: 404

This isn't an HTTP server. Apparently knew this at one point for the
DomeKey web site, but completely forgot.

Thanks to Ole (https://stackoverflow.com/users/312098/ole) and
RichieHindle (https://stackoverflow.com/users/21886/richiehindle) on
Stack Overflow for this question and answer that helped me figure it
out:
https://stackoverflow.com/questions/11223166/malformed-header-from-script-bad-header-http-1-1-302-found

Didn't notice this previously as I was using Lighttpd for testing, and
its FastCGI module must be more forgiving.