FastCgiRequest::path: Get path from `REQUEST_URI`
Use the `REQUEST_URI` param instead of `SCRIPT_NAME` to get the request
path.

Using `SCRIPT_NAME` caused issues with redirects.

If my FastCGI program has the file name `script.fcgi`, and I set up a
redirect from `/script` to `/script.fcgi`, then `path()` should return
`/script` when using that path.

In the above case, because we were using the `SCRIPT_NAME` param,
`path()` would return `/script.fcgi` instead.

This caused routes defined with a `RouteBuilder` to not match correctly.

Now that we're using `REQUEST_URI`, we need to trim the query string and
hash, which are included in the param value, since we only want the URI
path.